Fixed math.c build
