+++ src/utils/math.h
@@ -11,16 +11,6 @@
 
 #include <stdint.h>
 
-uint32_t roundUp(uint32_t numToRound, uint32_t multiple)
-{
-	if (multiple == 0)
-		return numToRound;
-
-	uint32_t remainder = numToRound % multiple;
-	if (remainder == 0)
-		return numToRound;
-
-	return numToRound + multiple - remainder;
-}
+uint32_t roundUp(uint32_t numToRound, uint32_t multiple);
 
 #endif // UTILS_MATH_H__
